+++ modules/planning/math/smoothing_spline/osqp_spline_1d_solver_test.cc
@@ -143,7 +143,7 @@ TEST(OsqpSpline1dSolver, three) {
   std::vector<double> ft_knots(21, 0.0);
 
   for (std::size_t i = 0; i < t_knots.size(); ++i) {
-    t_knots[i] = i * 0.1;
+    t_knots[i] = static_cast<double>(i) * 0.1;
     ft_knots[i] = t_knots[i] * slope + intercept;
   }
